fix: title shortitle
