--- targets/frontend/request.http
@@ -1,12 +1,0 @@
-## Requests available:
-
-###
-POST http://localhost:3001/api/actions/publish-all
-content-type: application/json
-
-{
-  "input": {
-    "source": "contributions"
-  }
-}
-
